Added some more OpCodes

+++ CHIP8.Infrastructure/Interfaces/ICHIP8Stack.cs
@@ -4,7 +4,7 @@ namespace CHIP8.Infrastructure.Interfaces
 {
     public interface ICHIP8Stack
     {
-        Byte Pop();
-        void Push(Byte value);
+        UInt16 Pop();
+        void Push(UInt16 value);
     }
 }
